requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
firebase-admin==6.2.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
openpyxl==3.1.2
tensorflow>=2.13.0
numpy>=1.24.3
Pillow>=10.0.0
pandas>=1.5.0
bcrypt==3.2.0
python-dotenv==0.19.0